backend/requirements.txt
# ReconSphere Python Dependencies
# Install with: pip install -r requirements.txt

pandas==2.1.4
openpyxl==3.1.2
xlrd==2.0.1
python-dotenv==1.0.0
google-generativeai==0.3.2